--- missingConfig.json
@@ -1,4 +1,0 @@
-{
-  "logLevel" : "INFO",
-  "userPrefsFilePath" : "preferences.json"
-}
