+++ src/main/java/seedu/address/logic/recommender/timing/exceptions/CommitmentClashException.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.recommender.timingrecommender.exceptions;
+package seedu.address.logic.recommender.timing.exceptions;
 
 /**
  * Represents an error where timing is not suitable for a module to fit into a time slot.
