+++ src/main/java/seedu/address/model/time/HourBlock.java
@@ -7,10 +7,10 @@ import java.util.Optional;
 import org.joda.time.Hours;
 import org.joda.time.LocalTime;
 
+import seedu.address.logic.recommender.timing.exceptions.CommitmentClashException;
 import seedu.address.model.commitment.Commitment;
 import seedu.address.model.time.exceptions.WrongTimeException;
 import seedu.address.model.time.util.TimeUtil;
-import seedu.address.model.timingrecommender.exceptions.CommitmentClashException;
 
 /**
  * Represents an hour timeslot in a Timetable.
